Move external school dto factory to provisioning module
