+++ apps/server/src/modules/provisioning/loggable/school-for-group-not-found.loggable.spec.ts
@@ -1,6 +1,5 @@
-import { externalSchoolDtoFactory } from '@testing/factory/external-school-dto.factory';
 import { ExternalGroupDto, ExternalSchoolDto } from '../dto';
-import { externalGroupDtoFactory } from '../testing';
+import { externalGroupDtoFactory, externalSchoolDtoFactory } from '../testing';
 import { SchoolForGroupNotFoundLoggable } from './school-for-group-not-found.loggable';
 
 describe('SchoolForGroupNotFoundLoggable', () => {
