+++ apps/server/src/modules/provisioning/testing/external-school-dto.factory.ts
@@ -1,6 +1,6 @@
-import { ExternalSchoolDto } from '@modules/provisioning/dto';
 import { ObjectId } from '@mikro-orm/mongodb';
 import { Factory } from 'fishery';
+import { ExternalSchoolDto } from '../dto';
 
 export const externalSchoolDtoFactory = Factory.define<ExternalSchoolDto, ExternalSchoolDto>(({ sequence }) => {
 	return {
