+++ apps/server/src/modules/provisioning/testing/index.ts
@@ -1,3 +1,4 @@
-export { externalUserDtoFactory } from './external-user-dto.factory';
 export { externalGroupDtoFactory } from './external-group-dto.factory';
 export { externalGroupUserDtoFactory } from './external-group-user-dto.factory';
+export { externalSchoolDtoFactory } from './external-school-dto.factory';
+export { externalUserDtoFactory } from './external-user-dto.factory';
